Pedestrian Detection Algorithm Learning
Adding Video Detection
Changing Image Detection from command line to IDE
